simulacion (alpha)
tengo que agregar una influencia de los valores sobre la simulación.
